Cleanups, split js wrapper & ts logic
