+++ packages/typegen/scripts/polkadot-types-from-defs.js
@@ -4,61 +4,21 @@
 // of the Apache-2.0 license. See the LICENSE file for details.
 /* eslint-disable @typescript-eslint/no-var-requires */
 
-// require('@babel/register')({
-//   extensions: ['.js', '.ts'],
-//   plugins: [
-//     ['module-resolver', {
-//       alias: {
-//         '^@polkadot/metadata(.*)': './packages/metadata/src\\1',
-//         '^@polkadot/typegen(.*)': './packages/typegen/src\\1',
-//         '^@polkadot/types(.*)': './packages/types/src\\1'
-//       }
-//     }]
-//   ]
-// });
+try {
+  require('./fromDefs');
+} catch (error) {
+  require('@babel/register')({
+    extensions: ['.js', '.ts'],
+    plugins: [
+      ['module-resolver', {
+        alias: {
+          '^@polkadot/metadata(.*)': './packages/metadata/src\\1',
+          '^@polkadot/typegen(.*)': './packages/typegen/src\\1',
+          '^@polkadot/types(.*)': './packages/types/src\\1'
+        }
+      }]
+    ]
+  });
 
-const path = require('path');
-const yargs = require('yargs');
-const substrateDefs = require('@polkadot/types/interfaces/definitions');
-const { generateInterfaceRegistry } = require('@polkadot/typegen/generateTypes/interfaceRegistry');
-const { generateTsDef } = require('@polkadot/typegen/generateTypes/tsDef');
-
-const { input, package } = yargs.strict().options({
-  input: {
-    description: 'The directory to use for the user definitions',
-    type: 'string',
-    required: true
-  },
-  package: {
-    description: 'The package name & path to use for the user types',
-    type: 'string',
-    required: true
-  }
-}).argv;
-
-const userDefs = require(path.join(process.cwd(), input, 'definitions.ts'));
-const userKeys = Object.keys(userDefs);
-const filteredBase = Object
-  .entries(substrateDefs)
-  .filter(([key]) => {
-    if (userKeys.includes(key)) {
-      console.warn(`Override found for ${key} in user types, ignoring in @polkadot/types`);
-
-      return false;
-    }
-
-    return true;
-  })
-  .reduce((defs, [key, value]) => {
-    defs[key] = value;
-
-    return defs;
-  }, {});
-
-const allDefs = {
-  '@polkadot/types/interfaces': filteredBase,
-  [package]: userDefs
-};
-
-generateTsDef(allDefs, path.join(process.cwd(), input), package);
-generateInterfaceRegistry(allDefs, path.join(process.cwd(), input, 'interfaceRegistry.ts'));
+  require('../sr/fromDefs.ts');
+}
